formatando os diretorios e apagando um metodo que nao estava sendo utilizado
