+++ bin/estrutura/threads/Reader.class
