+++ timeLog.txt
@@ -1,0 +1,101 @@
+(Readers: 0 | Writers: 100) >> Time: 187ms
+(Readers: 1 | Writers: 99) >> Time: 180ms
+(Readers: 2 | Writers: 98) >> Time: 180ms
+(Readers: 3 | Writers: 97) >> Time: 183ms
+(Readers: 4 | Writers: 96) >> Time: 178ms
+(Readers: 5 | Writers: 95) >> Time: 180ms
+(Readers: 6 | Writers: 94) >> Time: 180ms
+(Readers: 7 | Writers: 93) >> Time: 180ms
+(Readers: 8 | Writers: 92) >> Time: 180ms
+(Readers: 9 | Writers: 91) >> Time: 180ms
+(Readers: 10 | Writers: 90) >> Time: 181ms
+(Readers: 11 | Writers: 89) >> Time: 178ms
+(Readers: 12 | Writers: 88) >> Time: 178ms
+(Readers: 13 | Writers: 87) >> Time: 178ms
+(Readers: 14 | Writers: 86) >> Time: 178ms
+(Readers: 15 | Writers: 85) >> Time: 177ms
+(Readers: 16 | Writers: 84) >> Time: 178ms
+(Readers: 17 | Writers: 83) >> Time: 179ms
+(Readers: 18 | Writers: 82) >> Time: 178ms
+(Readers: 19 | Writers: 81) >> Time: 180ms
+(Readers: 20 | Writers: 80) >> Time: 178ms
+(Readers: 21 | Writers: 79) >> Time: 179ms
+(Readers: 22 | Writers: 78) >> Time: 179ms
+(Readers: 23 | Writers: 77) >> Time: 178ms
+(Readers: 24 | Writers: 76) >> Time: 179ms
+(Readers: 25 | Writers: 75) >> Time: 179ms
+(Readers: 26 | Writers: 74) >> Time: 183ms
+(Readers: 27 | Writers: 73) >> Time: 186ms
+(Readers: 28 | Writers: 72) >> Time: 184ms
+(Readers: 29 | Writers: 71) >> Time: 190ms
+(Readers: 30 | Writers: 70) >> Time: 192ms
+(Readers: 31 | Writers: 69) >> Time: 187ms
+(Readers: 32 | Writers: 68) >> Time: 183ms
+(Readers: 33 | Writers: 67) >> Time: 183ms
+(Readers: 34 | Writers: 66) >> Time: 183ms
+(Readers: 35 | Writers: 65) >> Time: 185ms
+(Readers: 36 | Writers: 64) >> Time: 183ms
+(Readers: 37 | Writers: 63) >> Time: 191ms
+(Readers: 38 | Writers: 62) >> Time: 182ms
+(Readers: 39 | Writers: 61) >> Time: 180ms
+(Readers: 40 | Writers: 60) >> Time: 183ms
+(Readers: 41 | Writers: 59) >> Time: 184ms
+(Readers: 42 | Writers: 58) >> Time: 182ms
+(Readers: 43 | Writers: 57) >> Time: 186ms
+(Readers: 44 | Writers: 56) >> Time: 181ms
+(Readers: 45 | Writers: 55) >> Time: 181ms
+(Readers: 46 | Writers: 54) >> Time: 179ms
+(Readers: 47 | Writers: 53) >> Time: 184ms
+(Readers: 48 | Writers: 52) >> Time: 178ms
+(Readers: 49 | Writers: 51) >> Time: 184ms
+(Readers: 50 | Writers: 50) >> Time: 180ms
+(Readers: 51 | Writers: 49) >> Time: 178ms
+(Readers: 52 | Writers: 48) >> Time: 182ms
+(Readers: 53 | Writers: 47) >> Time: 183ms
+(Readers: 54 | Writers: 46) >> Time: 180ms
+(Readers: 55 | Writers: 45) >> Time: 180ms
+(Readers: 56 | Writers: 44) >> Time: 184ms
+(Readers: 57 | Writers: 43) >> Time: 179ms
+(Readers: 58 | Writers: 42) >> Time: 189ms
+(Readers: 59 | Writers: 41) >> Time: 185ms
+(Readers: 60 | Writers: 40) >> Time: 184ms
+(Readers: 61 | Writers: 39) >> Time: 187ms
+(Readers: 62 | Writers: 38) >> Time: 185ms
+(Readers: 63 | Writers: 37) >> Time: 180ms
+(Readers: 64 | Writers: 36) >> Time: 185ms
+(Readers: 65 | Writers: 35) >> Time: 185ms
+(Readers: 66 | Writers: 34) >> Time: 180ms
+(Readers: 67 | Writers: 33) >> Time: 183ms
+(Readers: 68 | Writers: 32) >> Time: 181ms
+(Readers: 69 | Writers: 31) >> Time: 181ms
+(Readers: 70 | Writers: 30) >> Time: 176ms
+(Readers: 71 | Writers: 29) >> Time: 181ms
+(Readers: 72 | Writers: 28) >> Time: 185ms
+(Readers: 73 | Writers: 27) >> Time: 185ms
+(Readers: 74 | Writers: 26) >> Time: 184ms
+(Readers: 75 | Writers: 25) >> Time: 181ms
+(Readers: 76 | Writers: 24) >> Time: 187ms
+(Readers: 77 | Writers: 23) >> Time: 185ms
+(Readers: 78 | Writers: 22) >> Time: 178ms
+(Readers: 79 | Writers: 21) >> Time: 178ms
+(Readers: 80 | Writers: 20) >> Time: 182ms
+(Readers: 81 | Writers: 19) >> Time: 180ms
+(Readers: 82 | Writers: 18) >> Time: 181ms
+(Readers: 83 | Writers: 17) >> Time: 179ms
+(Readers: 84 | Writers: 16) >> Time: 174ms
+(Readers: 85 | Writers: 15) >> Time: 177ms
+(Readers: 86 | Writers: 14) >> Time: 180ms
+(Readers: 87 | Writers: 13) >> Time: 178ms
+(Readers: 88 | Writers: 12) >> Time: 178ms
+(Readers: 89 | Writers: 11) >> Time: 181ms
+(Readers: 90 | Writers: 10) >> Time: 179ms
+(Readers: 91 | Writers: 9) >> Time: 182ms
+(Readers: 92 | Writers: 8) >> Time: 177ms
+(Readers: 93 | Writers: 7) >> Time: 180ms
+(Readers: 94 | Writers: 6) >> Time: 177ms
+(Readers: 95 | Writers: 5) >> Time: 178ms
+(Readers: 96 | Writers: 4) >> Time: 184ms
+(Readers: 97 | Writers: 3) >> Time: 186ms
+(Readers: 98 | Writers: 2) >> Time: 180ms
+(Readers: 99 | Writers: 1) >> Time: 191ms
+(Readers: 100 | Writers: 0) >> Time: 186ms
